Stripe: Add support for idempotency keys
Stripe [has the ability to use an idempotency key][docs] to ensure that
a request is idempotent. This is very useful for things like retrying
a charge when Stripe returns an intermittent failure, as the request
will only be processed _once_ on Stripe's end even if the first time the
request was attempted it was actually processed.

[docs]: https://stripe.com/docs/api#idempotent_requests

Closes #1650.