Ezic: Add support for void
The void operation may or may not be supported by the underlying
processor sitting beneath Ezic.  That being said, some customers want to
utilize it if they can.

I'd like to have a test for a successful void, but the Ezic sandbox
doesn't seem to like it.  It ought to work in production though for many
processors.  We shall soon see.