CenPOS: Use ProcessCreditCard action
We were using the ProcessCard action rather than ProcessCreditCard.

According to the CenPOS docs:

ProcessCreditCard – Use for credit card transactions.
ProcessCard – Used to determine whether a card is credit, debit, or
commercial.

And:

ProcessCreditCard should always be used for card not present
transactions.

CenPOS support confirmed that this is the right action to use.